requirements.txt
numpy==1.24.3
requests
flask-socketio==5.5.1
annotated-types==0.7.0
anyio==4.8.0
appnope
archspec
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens
async-lru==2.0.4
attrs==25.1.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
boltons
Brotli
certifi
cffi
charset-normalizer
comm
coverage==7.6.10
cryptography
cycler==0.12.1
debugpy
decorator
defusedxml==0.7.1
distro
executing
fastjsonschema==2.21.1
filelock==3.17.0
flake8==7.1.1
fonttools==4.55.6
fqdn==1.5.1
frozendict
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna
ipykernel
ipython
ipywidgets==8.1.5
isoduration==20.11.0
jedi
Jinja2==3.1.5
json5==0.10.0
jsonpatch
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client
jupyter_core
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.1.0
mock==5.1.0
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio
networkx==3.2.1
nose==1.3.7
notebook==7.3.2
notebook_shim==0.2.4
ollama==0.4.7
opencv-python==4.8.0.74
overrides==7.7.0
packaging
pandas==2.2.3
pandocfilters==1.5.1
parso
pexpect
pillow==11.1.0
pip-conflict-checker==0.6.0
pipdeptree==2.24.0
platformdirs
pluggy
prometheus_client==0.21.1
prompt-toolkit
psutil
ptyprocess
pure-eval
py-cpuinfo==9.0.0
pycodestyle==2.12.1
pycosat
pycparser
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.2.1
PySocks
python-dateutil
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2024.2
PyYAML==6.0.2
pyzmq
referencing==0.36.2
requests
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
ruamel.yaml
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==72.1.0
six
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.2

stack-data
sympy==1.13.1
terminado==0.18.1
tinycss2==1.4.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado
tqdm
traitlets
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.71
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3
wcwidth
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.43.0
widgetsnbextension==4.0.13
flask-cors==4.0.0
groq
pyttsx3==2.90
zstandard